moved all component definitions to new XML style definitions
