+++ Showcases/geo_voyage/app/src/main/components/GrabbableNoRotation.xml
@@ -1,0 +1,13 @@
+<ComponentSchema packageName="com.meta.pixelandtexel.geovoyage">
+    <Component name="GrabbableNoRotation">
+        <Description>
+Similar to the Grabbable component, but applies zero rotation. To be used in conjunction with the
+Spinnable component (which instead handles the rotation).
+
+@param enabled Whether or not the grabbable functionality is enabled
+@param isGrabbed Used by system to track component state. You should not set this manually, and always use the default value.
+        </Description>
+        <BooleanAttribute name="enabled" defaultValue="true" />
+        <BooleanAttribute name="isGrabbed" defaultValue="false" />
+    </Component>
+</ComponentSchema>
