+++ Showcases/geo_voyage/app/src/main/components/Spin.xml
@@ -1,0 +1,12 @@
+<ComponentSchema packageName="com.meta.pixelandtexel.geovoyage">
+    <Component name="Spin">
+        <Description>
+A simple component to steadily rotate an entity around an axis.
+
+@param speed Degrees per second.
+@param axis The vector direction around which to rotate the object.
+        </Description>
+        <FloatAttribute name="speed" defaultValue="15f" />
+        <Vector3Attribute name="axis" defaultValue="0f, 1f, 0f" />
+    </Component>
+</ComponentSchema>
