+++ Showcases/geo_voyage/app/src/main/components/Spinnable.xml
@@ -1,0 +1,19 @@
+<ComponentSchema packageName="com.meta.pixelandtexel.geovoyage">
+    <Component name="Spinnable">
+        <Description>
+A component which allows the entity to be rotated around its yaw axis by the user, or pitched
+up and down relative to the user's location.
+
+@param size The size measurement of the entity with which to calculate the yaw rotation.
+@param drag How fast the entity rotation decelerates when the user stops manipulating.
+@param isGrabbed Used by system to track component state. You should not set this manually, and always use the default value.
+@param isSpinning Used by system to track component state. You should not set this manually, and always use the default value.
+@param pitch Used by system to track component state. You should not set this manually, and always use the default value.
+        </Description>
+        <FloatAttribute name="size" defaultValue="1f" />
+        <FloatAttribute name="drag" defaultValue="1f" />
+        <BooleanAttribute name="isGrabbed" defaultValue="false" />
+        <BooleanAttribute name="isSpinning" defaultValue="false" />
+        <FloatAttribute name="pitch" defaultValue="0f" />
+    </Component>
+</ComponentSchema>
