+++ Showcases/geo_voyage/app/src/main/components/Tether.xml
@@ -1,0 +1,19 @@
+<ComponentSchema packageName="com.meta.pixelandtexel.geovoyage">
+    <Component name="Tether">
+        <Description>
+A component which tethers one entity to another. The position can be adjusted, and the rotation
+is relative the the owning Entity's position and the user's head position.
+
+@param target Entity to which this owning Entity is tethered to.
+@param rotateAround How many degrees to rotate around the the entity, relative to the user.
+@param yawOffset After rotating around the entity, how much additional yaw to apply.
+@param zOffset Z position offset.
+@param yOffset Y position offset.
+        </Description>
+        <EntityAttribute name="target" />
+        <FloatAttribute name="rotateAround" defaultValue="90f" />
+        <FloatAttribute name="yawOffset" defaultValue="0f" />
+        <FloatAttribute name="zOffset" defaultValue="1f" />
+        <FloatAttribute name="yOffset" defaultValue="0f" />
+    </Component>
+</ComponentSchema>
